+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+* TBD:
+	* Follow standardization of query stringification as done by the lnurl-node module. The query object will now be sorted alphabetically by key before being stringified and used as the payload when signing.
 * v0.1.1:
 	* More examples and details about the library
 * v0.1.0: